+++ packages/core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [2.27.0](https://github.com/juntossomosmais/atomium/compare/atomium-v2.26.0...atomium-v2.27.0) (2024-11-25)
+
+
+### Features
+
+* **atom-steps-modal:** initial step feature ([#627](https://github.com/juntossomosmais/atomium/issues/627)) ([13965e5](https://github.com/juntossomosmais/atomium/commit/13965e584c77feec3704e50e046ed5dd6b6ec2cf))
+
+## [2.26.0](https://github.com/juntossomosmais/atomium/compare/atomium-v2.25.0...atomium-v2.26.0) (2024-11-19)
+
+
+### Features
+
+* **button:** add custom color to color prop ([#624](https://github.com/juntossomosmais/atomium/issues/624)) ([67ebf03](https://github.com/juntossomosmais/atomium/commit/67ebf037864df356fb951ff456425ec17fca9962))
+* **icons:** add new discount-tag icon ([#622](https://github.com/juntossomosmais/atomium/issues/622)) ([3989741](https://github.com/juntossomosmais/atomium/commit/39897417e4a6488e6e527e92e21bec0612a147dc))
+
 ## [2.25.0](https://github.com/juntossomosmais/atomium/compare/atomium-v2.24.3...atomium-v2.25.0) (2024-11-12)